refactor: align backend with eslint
