+++ moohaar-backend/.eslintrc.json
@@ -1,5 +1,18 @@
 {
   "extends": ["airbnb-base", "prettier"],
-  "env": { "browser": true, "es2022": true }
+  "env": { "node": true, "es2022": true },
+  "parserOptions": {
+    "ecmaVersion": "latest",
+    "sourceType": "module"
+  },
+  "overrides": [
+    {
+      "files": ["**/__tests__/**/*.js"],
+      "env": { "jest": true }
+    }
+  ],
+  "rules": {
+    "import/prefer-default-export": "error"
+  }
 }
