+++ moohaar-backend/package.json
@@ -24,7 +24,8 @@
     "mongoose": "^8.3.1",
     "multer": "^1.4.5-lts.1",
     "unzipper": "^0.10.11",
-    "winston": "^3.11.0"
+    "winston": "^3.11.0",
+    "joi": "^17.12.0"
   },
   "devDependencies": {
     "eslint": "^8.57.0",
