+++ moohaar-backend/src/controllers/theme.controller.js
@@ -3,15 +3,15 @@ import multer from 'multer';
 import path from 'path';
 import fs from 'fs/promises';
 
-import Theme from '../models/theme.model.js';
-import engine from '../services/liquid.service.js';
-import { auth, authorizeAdmin } from '../middleware/auth.middleware.js';
-import logger from '../utils/logger.js';
+import Theme from '../models/theme.model';
+import engine from '../services/liquid.service';
+import { auth, authorizeAdmin } from '../middleware/auth.middleware';
+import logger from '../utils/logger';
 
 // Utility for safe extraction and scanning of theme archives
-import sanitizeAndUnzip, {
+import unzipTheme, {
   MaliciousContentError,
-} from '../utils/unzip.util.mjs';
+} from '../utils/unzip.util';
 
 // Multer configuration: store uploads in configured path with size and type checks
 const upload = multer({
@@ -40,7 +40,7 @@ router.post(
 
       // Extract uploaded ZIP safely then remove the temporary archive
       try {
-        await sanitizeAndUnzip(req.file.path, destPath);
+        await unzipTheme(req.file.path, destPath);
       } catch (err) {
         await fs.unlink(req.file.path);
         if (err instanceof MaliciousContentError) {
@@ -188,7 +188,7 @@ router.put(
       // Remove existing directory and extract new one
       await fs.rm(themeDir, { recursive: true, force: true });
       try {
-        await sanitizeAndUnzip(req.file.path, themeDir);
+        await unzipTheme(req.file.path, themeDir);
       } catch (err) {
         await fs.unlink(req.file.path);
         if (err instanceof MaliciousContentError) {
