+++ moohaar-backend/src/routes/health.routes.js
@@ -1,5 +1,5 @@
 import { Router } from 'express';
-import { healthCheck } from '../controllers/health.controller.js';
+import healthCheck from '../controllers/health.controller';
 
 const router = Router();
