+++ moohaar-backend/src/services/db.service.js
@@ -1,12 +1,13 @@
 import mongoose from 'mongoose';
-import config from '../config/index.js';
+import config from '../config/index';
+import logger from '../utils/logger';
 
 const connectDB = async () => {
   try {
     await mongoose.connect(config.MONGODB_URI);
-    console.log('MongoDB connected');
+    logger.info('MongoDB connected');
   } catch (err) {
-    console.error('MongoDB connection error', err);
+    logger.error({ message: 'MongoDB connection error', error: err });
     throw err;
   }
 };
